+++ src/gitbook_downloader/gui/web/assets/swimlanesDiagram-VR7AAH4N-wfXrXEWb.js
@@ -1,8 +1,0 @@
-import{c as e,s as r}from"./flowDiagram-HODETNUW-DoJdNeAH.js";import{_ as a}from"./index-B7eRdI7e.js";import"./chunk-5VM5RSS4-D86M9B9l.js";import"./chunk-XXDRQBXY-BQCxfSXH.js";import"./chunk-POPQ4Y6H-BTSWY6Ul.js";import"./chunk-F27PBJKO-BLZv7HAs.js";import"./channel-BvSxQuV3.js";var o=a(t=>`${r(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,f=e({defaultLayout:"swimlane",styles:s});export{f as diagram};